refactor: domain 패키지 분리
1. domain.User
2. domain.Article
