+++ src/main/java/techcourse/myblog/domain/User/UserException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.domain.User;
 
 public class UserException extends RuntimeException {
