+++ src/main/java/techcourse/myblog/dto/UserDto.java
@@ -1,6 +1,6 @@
 package techcourse.myblog.dto;
 
-import techcourse.myblog.domain.User;
+import techcourse.myblog.domain.User.User;
 
 public class UserDto {
     private String name;
