+++ src/main/java/techcourse/myblog/repository/ArticleRepository.java
@@ -1,7 +1,7 @@
 package techcourse.myblog.repository;
 
 import org.springframework.data.repository.CrudRepository;
-import techcourse.myblog.domain.Article;
+import techcourse.myblog.domain.Article.Article;
 
 public interface ArticleRepository extends CrudRepository<Article, Long> {
