+++ src/main/java/techcourse/myblog/service/ArticleService.java
@@ -2,8 +2,8 @@ package techcourse.myblog.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import techcourse.myblog.domain.Article;
-import techcourse.myblog.domain.ArticleException;
+import techcourse.myblog.domain.Article.Article;
+import techcourse.myblog.domain.Article.ArticleException;
 import techcourse.myblog.dto.ArticleDto;
 import techcourse.myblog.repository.ArticleRepository;
