+++ src/test/java/techcourse/myblog/repository/ArticleRepositoryTest.java
@@ -8,7 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.context.junit4.SpringRunner;
-import techcourse.myblog.domain.Article;
+import techcourse.myblog.domain.Article.Article;
 import techcourse.myblog.dto.ArticleDto;
 
 import java.util.ArrayList;
