+++ src/test/java/techcourse/myblog/repository/UserRepositoryTest.java
@@ -5,9 +5,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import techcourse.myblog.domain.User;
-import techcourse.myblog.domain.UserEmail;
-import techcourse.myblog.domain.UserException;
+import techcourse.myblog.domain.User.User;
+import techcourse.myblog.domain.User.UserEmail;
+import techcourse.myblog.domain.User.UserException;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
