+++ src/test/java/techcourse/myblog/service/ArticleServiceTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import techcourse.myblog.domain.Article;
+import techcourse.myblog.domain.Article.Article;
 
 import java.util.List;
