+++ src/test/java/techcourse/myblog/service/LoginServiceTest.java
@@ -4,8 +4,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import techcourse.myblog.domain.User;
-import techcourse.myblog.domain.UserException;
+import techcourse.myblog.domain.User.User;
+import techcourse.myblog.domain.User.UserException;
 import techcourse.myblog.dto.UserDto;
 import techcourse.myblog.repository.UserRepository;
