+++ src/test/java/techcourse/myblog/web/ArticleControllerTestClient.java
@@ -11,7 +11,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.web.reactive.function.BodyInserters;
-import techcourse.myblog.domain.Article;
+import techcourse.myblog.domain.Article.Article;
 import techcourse.myblog.repository.ArticleRepository;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
