+++ src/test/java/techcourse/myblog/web/LoginControllerTest.java
@@ -10,7 +10,7 @@ import org.springframework.http.MediaType;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.reactive.server.WebTestClient;
 import org.springframework.web.reactive.function.BodyInserters;
-import techcourse.myblog.domain.User;
+import techcourse.myblog.domain.User.User;
 import techcourse.myblog.repository.UserRepository;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
